feat: export IOR through pipe
This makes it possible to get the tasks IOR without using CORBA name
service or messing with simultaneous file accesses.

The task would get the pipe write side from an external argument, and publish
sequential messages in the format of "task_name ior". Before closing the
pipe, a "DONE" message is emitted, signaling the end of transmission to
the pipe reader.